# docs/swagger.yaml
definitions:
...
  commands.BattleshipOptions:
    properties:
      time_per_turn:
        description: in seconds
        type: integer
      time_place_ship:
        description: in seconds
        type: integer
      who_go_first:
        description: '0: random, 1: player1, 2: player2'
        type: integer
    type: object
  commands.JoinRoomCommand:
    properties:
      name:
        type: string
      player_id:
        type: string
      room_id:
        type: string
      user_id:
        type: string
    required:
    - name
    - room_id
    type: object
...
  commands.PlayerCreateRoomCommand:
    properties:
      name:
        type: string
      options:
        $ref: '#/definitions/commands.BattleshipOptions'
      user_id:
        type: string
    required:
    - name
    type: object
...
  commands.SetWhoWinCommand:
    properties:
      player_id:
        type: string
    type: object
  commands.UpdateRoomStatusCommand:
    properties:
      room_id:
        type: string
      status:
        type: string
    required:
    - room_id
    - status
    type: object
...
  dtos.BattleshipOptionsDTO:
    properties:
      id:
        type: string
      room_id:
        description: UUID of the room
        type: string
      start_place_at:
        description: when the player can start placing ships
        type: string
      time_per_turn:
        description: in seconds
        type: integer
      time_place_ship:
        description: in seconds
        type: integer
      who_go_first:
        description: '0: random, 1: player1, 2: player2'
        type: integer
    type: object
  dtos.PlayerDTO:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  dtos.RoomDTO:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_ended:
        type: boolean
      options:
        $ref: '#/definitions/dtos.BattleshipOptionsDTO'
      status:
        type: string
      turn:
        type: integer
      updated_at:
        type: string
      who_win:
        type: integer
    type: object
  dtos.RoomPlayerDTO:
    properties:
      disconnected_at:
        type: string
      is_disconnected:
        type: boolean
      is_host:
        type: boolean
      is_ready:
        type: boolean
      me:
        type: integer
      player:
        $ref: '#/definitions/dtos.PlayerDTO'
      player_id:
        type: string
      room:
        $ref: '#/definitions/dtos.RoomDTO'
      room_id:
        type: string
      status:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/room/{roomId}/set-who-win:
    put:
      consumes:
      - application/json
      description: Sets the winner of a room based on the player ID.
      parameters:
      - description: Set Who Win Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/commands.SetWhoWinCommand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.RoomDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Set Who Win
      tags:
      - Room
  /api/v1/room/player/create:
    post:
      consumes:
      - application/json
      description: Allows a player to create a new room with a name and user ID
      parameters:
      - description: Player Create Room Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/commands.PlayerCreateRoomCommand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.RoomPlayerDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Player creates a new room
      tags:
      - Room.Player
  /api/v1/room/player/join:
    post:
      consumes:
      - application/json
      description: Allows a player to join an existing room with a name, user ID,
      parameters:
      - description: Player Join Room
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/commands.JoinRoomCommand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.RoomPlayerDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Player joins a room
      tags:
      - Room.Player
  /api/v1/room/status:
    put:
      consumes:
      - application/json
      description: Allows updating the status of a room by its ID
      parameters:
      - description: Update Room Status Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/commands.UpdateRoomStatusCommand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.RoomDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update room status
      tags:
      - Room
swagger: "2.0"